fix- correção erros por conflito de tipagem int->long em classes

+++ projeto-tfg/src/main/kotlin/school/sptech/projetotfg/domain/cadastro/PessoaDeficiencia.kt
@@ -10,7 +10,7 @@ import java.time.LocalDateTime
 @Entity
 class PessoaDeficiencia(
     @field:Id @field:GeneratedValue(strategy = GenerationType.IDENTITY) private var idPessoaDeficiencia:Long = 0,
-    @field:Max(255) private var verificacao: Int, //255 é o valor maximo de TINYLong
+    @field:Max(255) private var verificacao: Long, //255 é o valor maximo de TINYLong
     @field:ManyToOne private var deficiencia: Deficiencia,
     @field:PastOrPresent private var dataCriacao: LocalDateTime,
     @field:PastOrPresent private var dataUltimaAtualizacao: LocalDateTime,
